fixed profile page
